Refactor desktop App.tsx
