+++ erun-ui/frontend/src/app/clipboard.ts
@@ -1,0 +1,45 @@
+export function isTerminalPasteTarget(terminalRoot: HTMLDivElement, target: EventTarget | null): boolean {
+  return target instanceof Node && terminalRoot.contains(target);
+}
+
+export function pastedImageFiles(event: ClipboardEvent): File[] {
+  const items = event.clipboardData?.items;
+  if (!items) {
+    return [];
+  }
+
+  const files: File[] = [];
+  for (const item of Array.from(items)) {
+    if (item.kind !== 'file' || !item.type.toLowerCase().startsWith('image/')) {
+      continue;
+    }
+    const file = item.getAsFile();
+    if (file) {
+      files.push(file);
+    }
+  }
+  return files;
+}
+
+export async function fileToBase64(file: File): Promise<string> {
+  const buffer = await file.arrayBuffer();
+  return bytesToBase64(new Uint8Array(buffer));
+}
+
+export function bytesToBase64(bytes: Uint8Array): string {
+  const chunkSize = 0x8000;
+  let binary = '';
+  for (let index = 0; index < bytes.length; index += chunkSize) {
+    binary += String.fromCharCode(...bytes.subarray(index, index + chunkSize));
+  }
+  return window.btoa(binary);
+}
+
+export function decodeBase64Bytes(value: string): Uint8Array {
+  const binary = window.atob(value);
+  const bytes = new Uint8Array(binary.length);
+  for (let index = 0; index < binary.length; index += 1) {
+    bytes[index] = binary.charCodeAt(index);
+  }
+  return bytes;
+}
