+++ erun-ui/frontend/src/app/errors.ts
@@ -1,0 +1,12 @@
+export function readError(error: unknown): string {
+  if (typeof error === 'string') {
+    return error;
+  }
+  if (error instanceof Error && typeof error.message === 'string') {
+    return error.message;
+  }
+  if (error && typeof error === 'object' && 'message' in error && typeof error.message === 'string') {
+    return error.message;
+  }
+  return 'Unexpected error';
+}
