+++ erun-ui/frontend/src/components/app/IconTooltip.tsx
@@ -1,0 +1,12 @@
+import * as React from 'react';
+
+import { Tooltip, TooltipContent, TooltipTrigger } from '@/components/ui/tooltip';
+
+export function IconTooltip({ label, children }: { label: string; children: React.ReactElement }): React.ReactElement {
+  return (
+    <Tooltip>
+      <TooltipTrigger asChild>{children}</TooltipTrigger>
+      <TooltipContent>{label}</TooltipContent>
+    </Tooltip>
+  );
+}
